Update footer.html
I'm suggesting the leading slash be removed from this src path. It is resulting in the logo file not being found in https://asfadmin.github.io/asf-data-search-manual/

I'm not sure why this is working in HyP3, or is it?

Not sure, but it seems to be a problem in the search manual.